+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.3 (unreleased)
 
+- Allow setting max upgrade points in the lobby
 - Stop timer after game ends in Last Man Standing mode
 - Tiebreak by percent health remaining, not total health remaining